+++ example/test.js
@@ -13,14 +13,14 @@ var m2 = function() {
   console.log('m2');
 };
 
-var uri = 'mongodb://10.166.12.140:27017/r2do';
+var uri = 'mongodb://127.0.0.1:27017/r2do';
 
 // Read the certificate authority
 var ca = [fs.readFileSync(__dirname + "/cert.crt")];
 var cert = fs.readFileSync(__dirname + "/cert.pem");
 var key = fs.readFileSync(__dirname + "/cert.key");
 
-var mongoUseTls   = true;
+var mongoUseTls   = false;
 var elasticUseTls = false;
 
 // Connect
@@ -37,9 +37,9 @@ db.connect(uri, mongoUseTls ? {
     checkServerIdentity:false // if is self signed we must no enable server identifiy
   }
 } : { 
-  user : "r2do",
-  pass : "bs8R0P<LN6Zj2j5j]soC<HL]v$b]XYAx",
-  authMechanism : "SCRAM-SHA-1"
+  ///user : "r2do",
+  //pass : "bs8R0P<LN6Zj2j5j]soC<HL]v$b]XYAx",
+  //authMechanism : "SCRAM-SHA-1"
 }).then(function() {
   // load models
   db.models('./example/models');
@@ -47,7 +47,7 @@ db.connect(uri, mongoUseTls ? {
   db.methods('./example/methods');
   db.enums('./example/enums');
   // enable elastic hosts
-  db.elasticHosts([ { host : '37.187.183.24', port : 10300, protocol : 'http' }, { host : '37.187.183.24', port : 10200, protocol : 'http' } ],
+  db.enableElasticsearch([ { host : '127.0.0.1', port : 9200, protocol : 'http' } ],
     elasticUseTls ? {
       ssl: {
         ca: fs.readFileSync(__dirname + "/cert.pem"),
@@ -55,18 +55,23 @@ db.connect(uri, mongoUseTls ? {
       }
     } : {});
   // enable redis here
-  db.enableRedis([ { host : '10.166.13.10', port : 10600 } ]);
+  db.enableRedis([ { host : '127.0.0.1', port : 6379 } ]);
   if (db.isReady(true)) {
     db.load().then(function() {
-      db.getRedis().add('aezeazeaz', { took: 4,
+      for(var i = 0; i < 20; i ++) {
+      db.getRedis().add('aezeazeaz_'+i+Math.random(), { took: 4,
   timed_out: false,
   _shards: { total: 5, successful: 5, failed: 0 },
   hits: { total: 0, max_score: null, hits: [] } });
+      }
+
       /*db.getRedis().get('A').then(function(success) {
         console.log('redis s =>', success);
       }).catch(function (error) {
         console.log('redis eeee =>', error);
       })*/
+      db.getRedis().remove('aezeazeaz', 'rererer');
+      db.getRedis().flush('*33*');
       console.log('load success');
       var account = db.getModel('Account');
       console.log('===== value ===== ');
@@ -84,11 +89,16 @@ db.connect(uri, mongoUseTls ? {
         });
         
         account.insert({ name : "aee", test : 'fsdfds' }).then(function(a) {
-        console.log('A =>', a);
+        //console.log('A =>', a);
         account.modify(a._id.toString(), { name : 'rezrezre' }).then(function(u) {
-          console.log('U =>', u);
-          account.read(u._id.toString()).then(function(g) {
+          //console.log('U =>', u);
+          account.test1(u._id.toString()).then(function(g) {
             console.log('G =>' , g);
+            account.test2(u._id.toString()).then(function(d) {
+              //console.log('T2 =>', d);
+            }).catch(function(e) {
+              console.log('T2 failed', e);
+            });
             /*account.destroy(g._id.toString()).then(function(d) {
               console.log('D =>', d);
             }).catch(function(e) {
@@ -97,6 +107,7 @@ db.connect(uri, mongoUseTls ? {
           }).catch(function(e) {
             console.log('get failed', e);
           });
+
         }).catch(function(e) {
           console.log('update failed', e);
         })
@@ -109,6 +120,7 @@ db.connect(uri, mongoUseTls ? {
     });
   }
 }).catch(function(error) {
+  console.log(error);
   if (db.isConnected()) {
     db.disconnect().then(function() {